+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Added K factor for BSGS
 - Added bPfile.c to generate a precalculated file
 - Remove unused files about keccak and sha3
+- Change Bloom filter limits and % of error from 0.001 to 0.00001 in bloomfilter.
 
 #Version 0.1.20210112 BSGS
 - Added mode BSGS this work with a file with uncompressed keys